backend/pyproject.toml
[project]
authors = [
    {name = "Rami", email = "rami.awar.ra@gmail.com"},
    {name = "Anthony", email = "antounmalkoun@gmail.com"},
]
requires-python = "<3.12,>=3.11"
dependencies = [
    "sqlalchemy-redshift @ git+https://github.com/dtwelvespacespvt/sqlalchemy-redshift.git@develop#egg=sqlalchemy-redshift",
    "redshift-connector>=2.0.909",
    "fastapi==0.105.0",
    "uvicorn[standard]<1.0.0,>=0.21.1",
    "psycopg2-binary<3.0.0,>=2.9.9",
    "pymysql<2.0.0,>=1.1.1",
    "cryptography<41.0.0,>=40.0.2",
    "sqlalchemy[asyncio]<3.0.0,>=2.0.27",
    "rapidfuzz<4.0.0,>=3.0.0",
    "psycopg[pool]<4.0.0,>=3.1.9",
    "pydantic[email]==2.9.*",
    "openai<2.0.0,>=1.11.0",
    "python-jose==3.5.0",
    "google-auth==2.40.3",
    "sqlalchemy-bigquery==1.15.0",
    "psycopg-binary==3.2.3",
    # newer python-multipart versions break packages, see pypi changelog
    "python-multipart>=0.0.9,<=0.0.12",
    "aiosqlite<1.0.0,>=0.20.0",
    "alembic<2.0.0,>=1.13.1",
    "pydantic-settings<3.0.0,>=2.2.1",
    "asyncpg<1.0.0,>=0.29.0",
    "jinja2<4.0.0,>=3.1.3",
    "pytest-asyncio<1.0.0,>=0.23.6",
    "langchain==0.3.*",
    "langgraph==0.2.*",
    "langsmith<1.0.0,>=0.1.57",
    "langchain-openai==0.2.*",
    "mirascope==1.6.*",
    "tenacity<9.0.0,>=8.2.3",
    "pandas<3.0.0,>=2.2.2",
    "pyreadstat<2.0.0,>=1.2.7",
    "sentry-sdk[fastapi]<3.0.0,>=2.8.0",
    "snowflake-sqlalchemy<2.0.0,>=1.6.1",
    "openpyxl<4.0.0,>=3.1.5",
    "deepeval<1.0.0,>=0.21.55",
    "pyodbc<6.0.0,>=5.1.0",
    "posthog<4.0.0,>=3.6.6",
    "langchain-core==0.3.*",
    "langchain-community==0.3.*",
    "langchain-postgres>=0.0.15",
]
name = "dataline"
version = "0.1.0"
description = ""
readme = "README.md"

[dependency-groups]
dev = [
    "black<25.0.0,>=24.3.0",
    "pre-commit<4.0.0,>=3.4.0",
    "pyinstaller==6.11.*",
    "pytest<9.0.0,>=8.0.2",
    "jupyterlab<5.0.0,>=4.0.5",
    "mypy<2.0.0,>=1.6.1",
    "cx-freeze<7.0.0,>=6.15.14",
    "types-pygments<3.0.0.0,>=2.17.0.20240310",
    "pytest-env<2.0.0,>=1.1.3",
    "pylint-pydantic<1.0.0,>=0.3.2",
]

[tool.ruff]
ignore = ["E501"]
line-length = 120

[tool.black]
line-length = 120

[tool.mypy]
plugins = "pydantic.mypy"

show_error_codes = true

disallow_any_unimported = false
disallow_any_expr = false
disallow_any_decorated = false
disallow_any_explicit = true
disallow_any_generics = true
disallow_subclassing_any = false

disallow_untyped_calls = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = false

no_implicit_optional = true
no_implicit_reexport = true
strict_optional = true

warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unused_configs = true
warn_return_any = false
warn_unreachable = true

exclude = [
    '^bruno/*$',  # TOML literal string (single-quotes, no escaping necessary)
    "^notebooks/*$",  # TOML basic string (double-quotes, backslash and other characters need escaping)
    "^scripts/*$",
    "^hooks/*$",
    "^notebooks/*$",
    "^__pycache__/*$",
    "^venv/*$",
    "^build/*$",
]

[[tool.mypy.overrides]]
module = [
    "async_asgi_testclient",
    # https://github.com/hephex/asyncache/issues/14
    "asyncache",
    "deepmerge",
]


[tool.pydantic-mypy]
init_forbid_extra = true
init_typed = true
warn_required_dynamic_aliases = true

[tool.pylint.messages_control]
max-line-length = 120
disable = ["missing-class-docstring", "missing-function-docstring", "invalid-name", "unspecified-encoding", "missing-module-docstring"]
